wip: organize code
