+++ packages/fullstack/examples/island/src/components/counter.tsx
@@ -1,3 +1,4 @@
+import "./client.css";
 import { useState } from "preact/hooks";
 
 export function Counter() {
