+++ packages/fullstack/examples/island/vite-island.ts
@@ -1,0 +1,5 @@
+import type { Plugin } from "vite";
+
+export function islandPlugin(): Plugin[] {
+  return [];
+}
